refactor(enricher): ♻️ improve logging
