+++ README.md
@@ -1,3 +1,5 @@
+# AudioKit
+
 ![Vercel Deploy](https://deploy-badge.vercel.app/vercel/audiokitai-epsilon-records) [![Netlify Status](https://api.netlify.com/api/v1/badges/65f87109-7c6a-4a84-8f91-cb2ea8fc48bf/deploy-status)](https://app.netlify.com/sites/audiokit/deploys)
 
 ## Developing
