+++ src/lib/server/enrichers/hubspot.ts
@@ -1,6 +1,6 @@
 import { db } from '../../db/index.js';
 import { artists } from '../../db/schema.js';
-import { getHubspotContact } from '../hubspot.js';
+import { getHubspotContact } from '../integrations/hubspot.js';
 import { eq } from 'drizzle-orm';
 import logger from '../../utils/logger.js';
 import { error } from '@sveltejs/kit';
