+++ src/lib/server/integrations/hubspot.ts
@@ -1,6 +1,6 @@
-import type { artistSchema } from '../schemas/artist.js';
+import type { artistSchema } from '../../schemas/artist.js';
 import type { z } from 'zod';
-import logger from '../utils/logger.js';
+import logger from '../../utils/logger.js';
 import { serializeError } from 'serialize-error';
 import { inspect } from 'util';
