+++ src/lib/server/integrations/musicfetch.ts
@@ -1,4 +1,4 @@
-import logger from '../utils/logger.js';
+import logger from '../../utils/logger.js';
 import { serializeError } from 'serialize-error';
 import { inspect } from 'util';
