+++ src/lib/server/integrations/soundcharts.ts
@@ -1,6 +1,6 @@
-import type { Track, TrackCollectionResponse } from '../types/track.js';
+import type { Track, TrackCollectionResponse } from '../../types/track.js';
 import { error } from '@sveltejs/kit';
-import logger from '../utils/logger.js';
+import logger from '../../utils/logger.js';
 import { serializeError } from 'serialize-error';
 import { inspect } from 'util';
